feat(auth): implement unified login signup
